EES-3828 Move EditableKeyStat components into more `pages/release/content/components` directory
